backend/requirements.txt
agentic-doc==0.0.24
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
openai==1.3.0
numpy==1.24.3
fpdf==1.7.2
PyJWT==2.10.1
bcrypt==4.3.0
psycopg2-binary==2.9.9
requests==2.31.0
pillow==10.1.0
opencv-python==4.8.1.78

# Production dependencies
gunicorn==21.2.0
gevent==23.9.1
sentry-sdk==1.40.4  # For error tracking
prometheus-client==0.19.0  # For metrics monitoring

# Let pip resolve the remaining dependencies automatically